Coffee_Shop_Sales_Data_Insights
Problem Statement

In the competitive coffee sales industry, businesses need to understand sales performance and customer behavior to make informed decisions. However, analyzing sales data is challenging due to it being scattered across various sources and lacking a cohesive visualization tool. The key challenge is to collect, transform, and present this data in a way that provides stakeholders with clear and actionable insights.

### Steps Taken to Complete the Project

1. **Data Gathering:**
   - Sales data was collected from multiple sources, including transaction dates, country of sale, customer information, and sales amounts. This data was consolidated into a single Excel file for ease of analysis.

2. **Data Cleaning and Transformation:**
   - The dataset was reviewed for inaccuracies, duplicates, and missing values, and cleaned accordingly. Excel functions like VLOOKUP, IF statements, and TEXT functions were used to categorize sales by month, country, and customer.

3. **Creating the Dashboard:**
   - An interactive dashboard was developed using pivot tables and charts to allow dynamic data exploration. It displayed total sales trends over time, geographic performance by country, and highlighted the top 5 customers.

4. **Customization and Interactivity:**
   - Slicers and timelines were integrated into the dashboard to enhance user experience, enabling users to filter data based on their preferences. The dashboard was designed to be visually appealing and easy to navigate.


The project aimed to provide a comprehensive solution for businesses in the coffee sales industry to leverage their sales data for strategic decision-making. The interactive dashboard offers valuable insights for sales teams, management, and other stakeholders.